More changes to correct integer casting and Format String syntax errors related to move to 64-bit architecture.

+++ Accordion2/Classes/GITree.m
@@ -130,7 +130,7 @@
         if ([fileManager fileExistsAtPath:[node absolutePath]]) {
             NSError *error = nil;
             if (![fileManager removeItemAtPath:[node absolutePath] error:&error]) {
-                DDLogError(@"Error removing report at index: %lu ; error = %@",index,[error userInfo]);
+                DDLogError(@"Error removing report at index: %lu ; error = %@",(unsigned long)index,[error userInfo]);
             }
         }
         [_sortedNodes removeObjectAtIndex:index];
